Add kv_cache_factory
